+++ CHANGELOG.md
@@ -12,7 +12,9 @@ This project adheres to [Semantic Versioning](http://semver.org/) and
 - top-level alias functions for constructors of the basic datatypes
 - top-level constant functions `RDF.true` and `RDF.false` for the two boolean 
   RDF.Literal values
-- `RDF.Numeric` with a list of all numeric datatypes
+- `RDF.Decimal` datatype for `xsd:decimal` literals
+- `RDF.Numeric` with a list of all numeric datatypes and shared functions for
+  all numeric literals
 - the logical operators and the Effective Boolean Value (EBV) coercion algorithm 
   from the XPath and SPARQL specs on `RDF.Boolean`
 - `RDF.Term.equal?/2` and `RDF.Term.equal_value?/2`